Move package name to io.sentry.kotlin.multiplatform
