+++ src/androidMain/kotlin/io/sentry/kotlin/multiplatform/SentryAndroid.kt
@@ -1,6 +1,7 @@
-package io.sentry
+package io.sentry.kotlin.multiplatform
 
 import android.content.Context
+import io.sentry.Sentry
 import io.sentry.android.core.SentryAndroid
 
 internal actual object SentryBridge {
@@ -30,7 +31,7 @@ object ContextProvider {
     private lateinit var resolveContext: ()-> Context
 
     fun init(resolveContext: () -> Context) {
-        this.resolveContext = resolveContext
+        ContextProvider.resolveContext = resolveContext
     }
 
     fun getContext() : Context {
