--- src/jsTest/kotlin/io/sentry/SentryTestsJS.kt
@@ -1,5 +1,0 @@
-package io.sentry
-
-class SentryTestsJS {
-
-}
