+++ src/jsTest/kotlin/io/sentry/kotlin/multiplatform/SentryTestsJS.kt
@@ -1,0 +1,5 @@
+package io.sentry.kotlin.multiplatform
+
+class SentryTestsJS {
+
+}
